requirements (4).txt
Flask
pyttsx3
google-api-python-client
google-auth
